Add files via upload
The Readme is one single run of all project files. 

HOW TO INSTALL:

Depending on your system get pycharm, which should come with a python 2 interpretatior. If not google for the download. 

Download all the files, ALL, otherwise weird things will happen. Open the main file (DOAWB.py) and the rest should be taken care of.